MGS-4925 Create writable directory for jekyll to write into
This is needed by the jenkins job, since the 'jekyll build' command
is run within a docker container under UID 1000, which does not have
permissions to create _site in the working directory.

The documentation states that you can change the UID of the jekyll user
at runtime with the JEKYLL_ID environment variable, however setting this
to the UID of the user running the docker image fails.